+++ addon/popup_road_properties.py
@@ -15,7 +15,7 @@ import bpy
 
 import json
 import os
-from . helpers import get_user_cross_sections_path, call_operator_deferred
+from . helpers import get_user_cross_sections_path, call_operator_deferred, round_float_property
 
 
 # Operator to save the current cross-section as a user preset
@@ -68,21 +68,22 @@ class DSC_OT_save_cross_section_preset(bpy.types.Operator):
             'road_split_lane_idx': props.road_split_lane_idx,
             'road_mark_line_length': props.road_mark_line_length,
             'road_mark_line_space': props.road_mark_line_space,
-            'height_curb': props.height_curb
+            'height_curb': round_float_property(props.height_curb)
         }
 
         # Add lane data in correct order
         for lane in props.lanes:
             data['sides'].append(lane.side)
-            data['widths_start'].append(lane.width_start)
-            data['widths_end'].append(lane.width_end)
+            data['widths_start'].append(round_float_property(lane.width_start))
+            data['widths_end'].append(round_float_property(lane.width_end))
             data['types'].append(lane.type)
             data['road_mark_types'].append(lane.road_mark_type)
             data['road_mark_weights'].append(lane.road_mark_weight)
-            data['road_mark_widths'].append(lane.road_mark_width)
+            data['road_mark_widths'].append(round_float_property(lane.road_mark_width))
             data['road_mark_colors'].append(lane.road_mark_color)
             data['guard_rails'].append(lane.guard_rail)
-            data['guard_rail_lateral_offsets'].append(lane.guard_rail_lateral_offset)
+            data['guard_rail_lateral_offsets'].append(
+                round_float_property(lane.guard_rail_lateral_offset))
 
         # Get user JSON file path
         user_json_path = get_user_cross_sections_path()